♻️ split conversation authority types by concern
