+++ libs/backend/server/conversations/main/src/prisma-conversation-query-repository.ts
@@ -4,7 +4,10 @@ import { ConversationLifecycles, ConversationModes, MessageRoles, MessageSources
 import { AgentThreadDeliveryKinds, type AgentThreadParentDelivery } from "@opencrane/backend/conversations/agent-threads";
 import { __EncodeConversationProjectionCursor } from "@opencrane/backend/conversations/projection";
 
-import { AgentThreadRunViewStates, PersonalAgentDirectoryStatuses, type AgentThreadMessageView, type AgentThreadRunView, type AgentThreadSnapshotView, type ConversationCaller, type ConversationCreationDirectory, type ConversationDetail, type ConversationMessageView, type ConversationSummary } from "./conversation-authority.types.js";
+import { AgentThreadRunViewStates, type AgentThreadMessageView, type AgentThreadRunView, type AgentThreadSnapshotView } from "./types/agent-thread-view.types.js";
+import type { ConversationCaller } from "./types/conversation-caller.types.js";
+import { PersonalAgentDirectoryStatuses, type ConversationCreationDirectory } from "./types/conversation-directory.types.js";
+import type { ConversationDetail, ConversationMessageView, ConversationSummary } from "./types/conversation-view.types.js";
 import type { ConversationCommandContext, ConversationQueryRepository } from "./prisma-conversation-query-repository.types.js";
 
 /** Maximum message rows returned by one participant-bound open operation. */
