+++ libs/backend/server/conversations/main/src/prisma-conversation-query-repository.types.ts
@@ -1,6 +1,9 @@
 import type { ConversationLifecycles, ConversationModes } from "@opencrane/models/conversations";
 
-import type { AgentThreadSnapshotView, ConversationCaller, ConversationCreationDirectory, ConversationDetail, ConversationMessageView, ConversationSummary } from "./conversation-authority.types.js";
+import type { AgentThreadSnapshotView } from "./types/agent-thread-view.types.js";
+import type { ConversationCaller } from "./types/conversation-caller.types.js";
+import type { ConversationCreationDirectory } from "./types/conversation-directory.types.js";
+import type { ConversationDetail, ConversationMessageView, ConversationSummary } from "./types/conversation-view.types.js";
 
 /**
  * The four stored facts the command decision needs, and nothing else.
