+++ libs/backend/server/conversations/main/src/prisma-conversation-unit-of-work.ts
@@ -5,7 +5,13 @@ import { Prisma, type PrismaClient } from "@prisma/client";
 import { ___DoWithTrace } from "@opencrane/backend/observability";
 import type { RunRetryAuthority } from "@opencrane/backend/agents/execution/runs";
 
-import type { AgentThreadSnapshotView, ConversationCaller, ConversationCreationDirectory, ConversationDetail, ConversationSummary, ConversationUnitOfWork, CreateConversationRequest, CreateConversationResult, MarkAgentThreadReadResult, MutateConversationResult, RetryConversationRunRequest, RetryConversationRunResult, SubmitConversationMessageRequest, SubmitConversationMessageResult } from "./conversation-authority.types.js";
+import type { AgentThreadSnapshotView } from "./types/agent-thread-view.types.js";
+import type { CreateConversationResult, MarkAgentThreadReadResult, MutateConversationResult, RetryConversationRunResult, SubmitConversationMessageResult } from "./types/conversation-authority-result.types.js";
+import type { ConversationCaller } from "./types/conversation-caller.types.js";
+import type { ConversationCreationDirectory } from "./types/conversation-directory.types.js";
+import type { CreateConversationRequest, RetryConversationRunRequest, SubmitConversationMessageRequest } from "./types/conversation-request.types.js";
+import type { ConversationUnitOfWork } from "./types/conversation-unit-of-work.types.js";
+import type { ConversationDetail, ConversationSummary } from "./types/conversation-view.types.js";
 import type { ConversationMessageAdmissionUnitOfWork } from "./conversation-message-admission.types.js";
 import { PrismaConversationMutationRepository } from "./prisma-conversation-mutation-repository.js";
 import type { ConversationMutationRepository } from "./prisma-conversation-mutation-repository.types.js";
